chore: extracted Paged proeprties
