+++ dhis-2/dhis-api/src/main/java/org/hisp/dhis/datavalue/DataValueChangelogQueryParams.java
@@ -33,9 +33,11 @@ import java.util.List;
 import org.hisp.dhis.category.CategoryOptionCombo;
 import org.hisp.dhis.common.OpenApi;
 import org.hisp.dhis.common.UID;
-import org.hisp.dhis.common.UrlParams;
+import org.hisp.dhis.common.input.Paged;
+import org.hisp.dhis.common.input.UrlParams;
 import org.hisp.dhis.dataelement.DataElement;
 import org.hisp.dhis.dataset.DataSet;
+import org.hisp.dhis.jsontree.Collapsed;
 import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.period.Period;
 
@@ -53,10 +55,7 @@ public record DataValueChangelogQueryParams(
     @OpenApi.Property({UID[].class, CategoryOptionCombo.class}) UID co, // COC
     @OpenApi.Property({UID[].class, CategoryOptionCombo.class}) UID cc, // AOC
     List<DataValueChangelogType> type,
-    Boolean skipPaging,
-    boolean paging,
-    int page,
-    int pageSize)
+    @Collapsed Paged paged)
     implements UrlParams {
 
   public static final DataValueChangelogQueryParams DEFAULT = ofType();
@@ -71,15 +70,6 @@ public record DataValueChangelogQueryParams(
         null,
         null,
         List.of(types),
-        null,
-        true,
-        1,
-        50);
-  }
-
-  @OpenApi.Ignore
-  public boolean isPaged() {
-    if (skipPaging != null) return !skipPaging;
-    return paging;
+        Paged.DEFAULT);
   }
 }
