+++ dhis-2/dhis-web-api/src/main/java/org/hisp/dhis/webapi/controller/MinMaxDataElementController.java
@@ -44,6 +44,7 @@ import org.hisp.dhis.common.Maturity;
 import org.hisp.dhis.common.OpenApi;
 import org.hisp.dhis.common.Pager;
 import org.hisp.dhis.common.UID;
+import org.hisp.dhis.common.input.Paged;
 import org.hisp.dhis.csv.CSV;
 import org.hisp.dhis.datavalue.DataValue;
 import org.hisp.dhis.feedback.BadRequestException;
@@ -103,9 +104,10 @@ public class MinMaxDataElementController {
 
     RootNode rootNode = NodeUtils.createMetadata();
 
-    if (query.isPaged()) {
+    Paged paged = query.paged();
+    if (paged.isPaged()) {
       int total = minMaxService.countMinMaxDataElements(query);
-      Pager pager = new Pager(query.page(), total, query.pageSize());
+      Pager pager = new Pager(paged.page(), total, paged.pageSize());
       rootNode.addChild(NodeUtils.createPager(pager));
     }
